Update patch set 1
Patch Set 1:

I want to run a build with this before I pull it. I see a lot of other devices using it, so I don't expect a problem, but I want to see it work first. May be a couple days before I have time due to work on cm-14.0, however.

Patch-set: 1