Update patch set 1
Patch Set 1: Code-Review-1

The more I think about this, the less I like flashing the bootloader with every nightly. I'd like to find a way to detect bootloader version and use that to detect if an update is needed or not, but I haven't found anything yet.

Patch-set: 1
Label: Code-Review=-1